[JBTM-3463] split up each service (hotel, flight and trip) so that they each run in their own JVMs,
