+++ rts/at/annotation/common-service/pom.xml
@@ -1,0 +1,19 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/maven-v4_0_0.xsd">
+  <modelVersion>4.0.0</modelVersion>
+  <parent>
+    <groupId>org.jboss.narayana.quickstart.rts</groupId>
+    <artifactId>annotation</artifactId>
+    <version>7.0.1.Final-SNAPSHOT</version>
+  </parent>
+  <groupId>org.jboss.narayana.quickstart.rts</groupId>
+  <artifactId>common-service</artifactId>
+  <packaging>jar</packaging>
+  <name>common-service</name>
+  <dependencies>
+    <dependency>
+      <groupId>com.fasterxml.jackson.core</groupId>
+      <artifactId>jackson-annotations</artifactId>
+    </dependency>
+  </dependencies>
+</project>
