+++ rts/at/annotation/common-service/src/main/java/io/narayana/sra/demo/service/BookingException.java
@@ -1,0 +1,16 @@
+package io.narayana.sra.demo.service;
+
+public class BookingException extends Exception {
+    int reason;
+
+    public BookingException(int reason, String message) {
+        super(message);
+
+        this.reason = reason;
+
+    }
+
+    public int getReason() {
+        return reason;
+    }
+}
