+++ rts/at/annotation/common-service/src/main/resources/META-INF/services/jakarta.ws.rs.ext.Providers
