+++ rts/at/annotation/flight-service/src/main/resources/application.properties
@@ -1,0 +1,1 @@
+quarkus.http.port=8084
