+++ rts/at/annotation/hotel-service/src/main/resources/META-INF/beans.xml
@@ -1,0 +1,5 @@
+<beans version="4.0" bean-discovery-mode="annotated"
+       xmlns="https://jakarta.ee/xml/ns/jakartaee"
+       xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+       xsi:schemaLocation="https://jakarta.ee/xml/ns/jakartaee https://jakarta.ee/xml/ns/jakartaee/beans_4_0.xsd">
+</beans>
