+++ rts/at/annotation/trip-service/src/main/resources/META-INF/services/jakarta.ws.rs.ext.Providers
@@ -1,0 +1,1 @@
+org.jboss.jbossts.star.client.ServerSRAFilter
