+++ rts/at/annotation/trip-service/src/main/resources/application.properties
@@ -1,0 +1,3 @@
+quarkus.http.port=8082
+hotel.quarkus.http.port=8083
+flight.quarkus.http.port=8084
