+++ rts/at/pom.xml
@@ -59,6 +59,6 @@
     <module>recovery</module>
     <module>jta-service</module>
     <module>undertow</module>
-    <module>sra</module>
+    <module>annotation</module>
   </modules>
 </project>
